requirements.txt
av
datasets
einops
fastapi
fire
gradio
huggingface_hub
imageio
jieba
lmdb
matplotlib
modelscope
ndjson
nltk
numpy==1.26.4
nuscenes_devkit
openai
opencv_contrib_python
opencv_python
packaging
pandas
peft
Pillow
psutil
pydantic
pyquaternion
pytorch_lightning
PyYAML
ray
Requests==2.32.4
rouge_chinese
safetensors
setuptools==76.0.0
sse_starlette
tiktoken
tqdm
transformers>=4.49.0
trl
typing_extensions
uvicorn
vllm>=0.7.3
